Add User-Agent to Nakadi requests
We would like to have an easy way to get some statistics around Fahrschein usage in general, and which versions, which JDK version, and via which HTTP client implementation specificially, in order to a) justify the time we're spending on the library, b) get more insights about builder behaviour. The header should be set by default (not overridable initially).

Format

    "Fahrschein/" [FahrscheinVersion] " (" [HttpClient] "; " [Java Version] ("; " [Additional Property])* ")"

Examples

    User-Agent: Fahrschein/0.22.0 (Simple; Java11)
    User-Agent: Fahrschein/0.22.0 (ApacheHttpClient; Java8)
    User-Agent: Fahrschein/0.22.0 (JavaNet; Java17; experimental-features=a,b,c)
    User-Agent: Fahrschein/0.22.0 (Spring; Java8; Kotlin)